backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
Pillow==10.0.1
torch==2.0.1
torchvision==0.15.2
ftfy==6.1.1
regex==2023.10.3
tqdm==4.66.1
clip @ git+https://github.com/openai/CLIP.git
transformers==4.35.0
accelerate==0.23.0